requirements.txt
streamlit==1.36.0
tensorflow-cpu==2.20.0
numpy==2.3.5
pillow==10.4.0
